requirements.txt
fastapi==0.114.1
uvicorn==0.30.6
sqlalchemy==2.0.34
psycopg2-binary==2.9.9
python-dotenv==1.0.1
alembic==1.12.1